requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
aiosqlite>=0.20.0
redis>=5.0.0
bcrypt>=4.0.0
